+++ scripts/coverage.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+
+coverage run --source=./obscurepy -m unittest discover -s ./ -t ./
+
+coverage report